+++ tasq/__init__.py
@@ -9,4 +9,4 @@ from tasq.remote.supervisor import (ZMQActorSupervisor,
                                     RedisActorSupervisor,
                                     Supervisors)
 
-__version__ = '1.1.7'
+__version__ = '1.1.8'